// src/app/shared/enum/portal.enum.ts
export const enum PortalEnum {
  immostreet = '1',
  homegate = '2',
  immoscout = '3',
  homech = '4',
  publimmo = '5',
  seLoger = '6',
  logicImmo = '7',
  newhomech = '8',
  previsite = '9',
  immobilierch = '10',
  luxuryEstate = '11',
  immoving = '12',
  urbanhome = '13',
  icasa = '14',
  shelterr = '15',
  locanto = '16',
  tatanka = '17',
  spiti360 = '18',
  futurhome = '19',
  websiteIdx = '20',
  justimmo = '21',
  repimmo = '22',
  petitesannoncesch = '23',
  immoNouvelListe = '24',
  scanpark = '25',
  helvethome = '26',
  customIdx = '27',
  begehungen = '28',
  jamesEdition = '29',
  lefigaro = '48',
  immowelt = '49',
  propertyPartners = '50',
  bazzileRealforce = '51',
  snpi = '52',
}